remove workaround when deleting resources
